Separate NDArrayTransformation
